smile: enhance backend for collector
The collector will have to deal with payload containing multiple
serialized events (as we want to leverage compression with back
references).

This introduces a new concept: SmileBucket, which is a collection of
events (JsonNode). The collector will deserialize the payload into
SmileEnvelopeEvents, each associated with one JsonNode in the SmileBucket.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>