smile: introduce SmileBucketEvent
The entire writer code deals with Events only. For instance,
when writing Sequence files, the value is a serialized Event.

To leverage compression on disk and in HDFS, we needed an extra level of
abstraction. SmileBucketEvent is a meta event, working as a thin wrapper around
a SmileBucket.

This will also simplify the collector code: we can stop dealing with
lists of events and instead carry SmileBucketEvents around.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>